+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,8 +1,8 @@
 package blackjack.domain.participant;
 
-import blackjack.domain.Card;
-import blackjack.domain.Deck;
-import blackjack.domain.Hand;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
+import blackjack.domain.card.Hand;
 import java.util.List;
 
 public class Dealer extends Participant {
