+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -2,9 +2,9 @@ package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.*;
 
-import blackjack.domain.Card;
-import blackjack.domain.Denomination;
-import blackjack.domain.Suit;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Denomination;
+import blackjack.domain.card.Suit;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
